Move printer method to formatter
