+++ lib/watson/formatters.rb
@@ -1,0 +1,4 @@
+module Watson::Formatters
+  autoload :BaseFormatter, 'watson/formatters/base_formatter'
+  autoload :DefaultFormatter, 'watson/formatters/default_formatter'
+end
