+++ lib/watson/formatters/default_formatter.rb
@@ -1,0 +1,4 @@
+module Watson::Formatters
+  class DefaultFormatter < BaseFormatter
+  end
+end
